- Updated Xcode project to build last commit's newest changes. - Fixed demo project window "build.settings". - Resolved most compiler warnings in Xcode. - Updated Corona *.lib dependency files from build #2842 to #2872.

+++ Source/RuntimeContext.cpp
@@ -183,7 +183,7 @@ RuntimeContext* RuntimeContext::GetInstanceBy(lua_State* luaStatePointer)
 
 int RuntimeContext::GetInstanceCount()
 {
-	return sRuntimeContextCollection.size();
+	return (int)sRuntimeContextCollection.size();
 }
 
 int RuntimeContext::OnCoronaEnterFrame(lua_State* luaStatePointer)
